Create .swift
This commit is in Swift for the purpose of making a folder structure for this common sequence locator. The Models folder contains the necessary models for the program to define the data structures and algorithms necessary for finding the common sequence. The Algorithms folder includes the classes LongestCommonSequence and SubsequenceLocator that use algorithms for finding the appropriate sequence. The Tests folder contains unit tests for models and algorithms to guarantee their reliability. The main.swift file serves as the entry point of the program, written in a coding language that's tailored for iOS mobile devices.